Revert "EK-108: by cszablew, fixed <img> inline css, renamed files to PascalCaseX"
This reverts commit 554a41c612fa141ae10763508a53e55a3ed63458.
