+++ gatsby_fp/src/components/fieldComponents/genInfoTable.js
@@ -28,6 +28,7 @@ export default function GenInfoTable({
                   <img
                     src={`data:image/jpeg;base64,${imgsrc}`}
                     className="bioimage"
+                    style={{ maxWidth: "150px" }}
                     alt=""
                   />
                 </td>
