+++ gatsby_fp/src/components/fieldComponents/intellCont.js
